Fixed Reach Router Moved file directories to be together. Kept Form component name as Post Form Make sure to do localhost:8888/#/routename
