+++ app/config/config.json
@@ -1,0 +1,13 @@
+{
+    "development": {
+        "username": "root",
+        "password": "",
+        "database": "burger",
+        "host": "127.0.0.1",
+        "dialect": "mysql"
+    },
+    "production": {
+        "use_env_variable": "JAWSDB_URL",
+        "dialect": "mysql"
+    }
+}
