--- config/connection.js
@@ -1,20 +1,0 @@
-// *********************************************************************************
-// CONNECTION.JS - THIS FILE INITIATES THE CONNECTION TO MYSQL
-// *********************************************************************************
-
-// Dependencies
-var Sequelize = require("sequelize");
-
-// Creates mySQL connection using Sequelize
-var sequelize = new Sequelize("starwars", "root", "", {
-  host: "localhost",
-  dialect: "mysql",
-  pool: {
-    max: 5,
-    min: 0,
-    idle: 10000
-  }
-});
-
-// Exports the connection for other files to use
-module.exports = sequelize;
